Fix bugs that prevented imports of more than 20 billion nodes from working, in environments where the amount of RAM available to the importer, was close to the amount of RAM required for the import.

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/cache/HeapIntArray.java
@@ -29,7 +29,7 @@ public class HeapIntArray extends HeapNumberArray<IntArray> implements IntArray
     private final int[] array;
     private final int defaultValue;
 
-    public HeapIntArray( int length, int defaultValue, int base )
+    public HeapIntArray( int length, int defaultValue, long base )
     {
         super( 4, base );
         this.defaultValue = defaultValue;
